Fix fragment meta to be 4 bytes for sequence numbers

+++ LoraGateway.Terminal/Services/Firmware/RandomLinearCoding/FragmentWithGenerator.cs
@@ -3,7 +3,7 @@
 public class FragmentWithGenerator
 {
     public byte UsedGenerator { get; set; }
-    public byte SequenceNumber { get; set; }
+    public UInt16 SequenceNumber { get; set; }
     public byte GenerationIndex { get; set; }
     public byte[] Fragment { get; set; }
